2.0.4-玉衡-Alpha-31 [修复]修复若干问题

+++ app/src/main/java/com/imcys/bilibilias/home/ui/adapter/PlayHistoryAdapter.kt
@@ -48,7 +48,7 @@ class PlayHistoryAdapter @Inject constructor() :
                 val bvid = if (getItem(position).history.bvid != "") {
                     getItem(position).history.bvid
                 } else {
-                    VideoNumConversion.toBvidOffline(getItem(position).history.oid.toLong())
+                    VideoNumConversion.toBvidOffline(getItem(position).history.oid)
                 }
                 AsVideoActivity.actionStart(holder.itemView.context, bvid)
             }
